+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Add join conditions to aggregate queries
+
 ## 0.5.0 - 2017-07-02
 ### Added
 - Soft deletes when model has `::$softDelete` property.