+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.5.2 - 2017-08-04
 ### Fixed
 - Validating values with multiple validation rules short-circuits on the first failed rule.